+++ hardware/nrf54l15clean/nrf54l15clean/libraries/Nrf54L15-Clean-Implementation/src/ble_nus.h
@@ -11,7 +11,7 @@ class BleNordicUart : public Stream {
  public:
   static constexpr size_t kRxBufferSize = 4096U;
   static constexpr size_t kTxBufferSize = 4096U;
-  static constexpr uint8_t kMaxPayloadLength = BleRadio::kCustomGattMaxValueLength;
+  static constexpr uint16_t kMaxPayloadLength = BleRadio::kCustomGattMaxValueLength;
 
   static const uint8_t kServiceUuid128[16];
   static const uint8_t kRxCharacteristicUuid128[16];
@@ -41,7 +41,7 @@ class BleNordicUart : public Stream {
 
   uint32_t rxDroppedBytes() const;
   uint32_t txDroppedBytes() const;
-  uint8_t maxPayloadLength() const;
+  uint16_t maxPayloadLength() const;
 
   void clear();
   void clearRx();
@@ -59,12 +59,12 @@ class BleNordicUart : public Stream {
 
  private:
   static void onRxWriteThunk(uint16_t valueHandle, const uint8_t* value,
-                             uint8_t valueLength, bool withResponse,
+                             uint16_t valueLength, bool withResponse,
                              void* context);
 
-  void onRxWrite(const uint8_t* value, uint8_t valueLength);
+  void onRxWrite(const uint8_t* value, uint16_t valueLength);
   bool queueNextNotification();
-  uint8_t notificationValueLimit() const;
+  uint16_t notificationValueLimit() const;
   size_t copyTxChunk(uint8_t* outChunk, size_t maxLength) const;
   void resetSessionState();
 
@@ -82,7 +82,7 @@ class BleNordicUart : public Stream {
   uint8_t rxBuffer_[kRxBufferSize];
   uint8_t txBuffer_[kTxBufferSize];
   uint8_t txChunk_[kMaxPayloadLength];
-  uint8_t txChunkLength_;
+  uint16_t txChunkLength_;
   uint32_t rxDroppedBytes_;
   uint32_t txDroppedBytes_;
   uint32_t debugNotificationSentCount_;
